Remove limitations on -v flag
While it doesn't make sense to have it with the -r option or occurring
alone (list command), it's not a problem -- it just doesn't do
anything.
Furthermore, being a bit more accepting with respect to this flag allows
it to be used in an alias, in order to have it always present without
typing it in by hand.